2/4

+++ app/Http/Middleware/QuizRepetition.php
@@ -19,7 +19,18 @@ class QuizRepetition
         $attempt = $request->attributes->get('attempt');
 
         if($attempt && $attempt->status_passed == "repetition"){
-            $attempt->delete();
+            $currentTime = now();
+            $endTime = $attempt->created_at->addMinutes(2);
+            if($currentTime >= $endTime){
+                $attempt->delete();
+            }else{
+                $remainingTime = $endTime->diffInMinutes($currentTime);
+                return response()->json([
+                    'status_code' => 403,
+                    'success' => false,
+                    'message' => 'There are '.$remainingTime.' minutes left to retake the quiz'
+                  ], 200);
+            }
         }
         return $next($request);
     }
